[BFN] divided classes Fan and  FanDrawer into 2 files
